fix(backend): harden auth, money handling, and scraping architecture
Upgrade authentication to fastapi-users 15.0.5 with pwdlib/argon2
hashing and PyJWT, replacing the deprecated python-jose/passlib stack.

Represent all monetary values as Decimal backed by NUMERIC(10,2)
columns (market_price, purchase_price) and store datetimes as aware
UTC via a custom UTCDateTime TypeDecorator, with an Alembic migration
(0002) that rebuilds the affected columns and recreates the partial
unique index on (game, external_id).

Refactor the scraping layer into a shared-client architecture:
- Register scrapers by game with process-wide singletons and a
  single-flight map so concurrent searches never duplicate requests
- Add a 24h TTL cache on card data and fan out multi-game searches
  with asyncio.gather, isolating each game's upsert in its own
  AsyncSession to avoid flush races
- Confine Playwright usage to the Yu-Gi-Oh! scraper behind a shared
  browser singleton with serialized page access
- Degrade to local results with structured logging when external
  sources fail, and close all clients/browser on app shutdown

Fix collection listing returning total = N^2 by counting over the
joined subquery without re-importing the base table, and apply an
explicit deterministic ORDER BY on list queries.

Add slowapi rate limiting (toggleable) for login, search, and price
refresh endpoints, and disable SQLite foreign keys during Alembic
batch rebuilds so cascading deletes cannot wipe data mid-migration.

+++ src/schemas/collection.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from datetime import datetime
+from decimal import Decimal
 
 from pydantic import BaseModel, Field
 
@@ -14,7 +15,7 @@ class CollectionAddRequest(BaseModel):
     condition: Condition = Condition.NEAR_MINT
     is_foil: bool = False
     language: str = Field(default="EN", max_length=10)
-    purchase_price: float | None = Field(default=None, ge=0)
+    purchase_price: Decimal | None = Field(default=None, ge=0)
 
 
 class CollectionUpdateRequest(BaseModel):
@@ -22,7 +23,7 @@ class CollectionUpdateRequest(BaseModel):
     condition: Condition | None = None
     is_foil: bool | None = None
     language: str | None = Field(default=None, max_length=10)
-    purchase_price: float | None = Field(default=None, ge=0)
+    purchase_price: Decimal | None = Field(default=None, ge=0)
 
 
 class CollectionItemResponse(BaseModel):
@@ -33,10 +34,10 @@ class CollectionItemResponse(BaseModel):
     condition: Condition
     is_foil: bool
     language: str
-    purchase_price: float | None
+    purchase_price: Decimal | None
     added_at: datetime
     card: CardResponse
-    total_value: float
+    total_value: Decimal
 
     model_config = {"from_attributes": True}
 
@@ -50,6 +51,6 @@ class CollectionListResult(BaseModel):
 
 
 class PortfolioValueResponse(BaseModel):
-    total_value: float
+    total_value: Decimal
     cards_count: int
     unique_cards: int
